Fixed code highlighting issues in preview mode
My solution was to import a client-side code highlighting library, which
from an efficiency stand point is bad, because their is already a code
highlighting library on the server. However, this was the only way I
could think to allow the new markdown editor to provide a preview of how
the code highlighting will look without having to wait 5-10 seconds each
time the user makes a change for the server to respond with the correct
highlighting.